KeyedDistributions

KeyedDistributions.KeyedDistribution

KeyedDistributions.KeyedDistribution

KeyedDistributions.KeyedDistribution

KeyedDistributions.KeyedSampleable

KeyedDistributions.KeyedSampleable

AxisKeys.axiskeys

KeyedDistributions.distribution

KeyedDistributions.KeyedDistribution — Type

KeyedDistribution(d<:Distributions.Distribution, keys::Tuple{Vararg{AbstractVector}})

Stores keys for each variate alongside the Distributions.Distribution d, supporting all of the common functions of a Distributions.Distribution. Common functions that return an AbstractArray, such as rand, will return a KeyedArray with keys derived from the Distributions.Distribution.

The type of keys is restricted to be consistent with AxisKeys.jl. The length of the keys tuple must be the number of dimensions, which is 1 for univariate and multivariate distributions, and 2 for matrix-variate distributions. The length of each key vector in must match the length along each dimension.

KeyedDistributions.KeyedDistribution — Method

KeyedDistribution(d::Distribution)

Constructs a KeyedDistribution using the keys of the first field stored in d, or if there are no keys, 1:n for the length n of each dimension.

KeyedDistributions.KeyedDistribution — Method

KeyedDistribution(d<:Distributions.Distribution, keys::AbstractVector)

Constructor for KeyedDistribution with one dimension of variates. The elements of keys correspond to the variates of the distribution.

KeyedDistributions.KeyedSampleable — Type

KeyedSampleable(d<:Distributions.Sampleable, keys::Tuple{Vararg{AbstractVector}})

Stores keys for each variate alongside the Distributions.Sampleable d, supporting all of the common functions of a Distributions.Sampleable. Common functions that return an AbstractArray, such as rand, will return a KeyedArray with keys derived from the Distributions.Sampleable.

The type of keys is restricted to be consistent with AxisKeys.jl. The length of the keys tuple must be the number of dimensions, which is 1 for univariate and multivariate distributions, and 2 for matrix-variate distributions. The length of each key vector in must match the length along each dimension.

KeyedDistributions.KeyedSampleable — Method

KeyedSampleable(d<:Distributions.Sampleable, keys::AbstractVector)

Constructor for KeyedSampleable with one dimension of variates. The elements of keys correspond to the variates of the distribution.

AxisKeys.axiskeys — Method

axiskeys(d::Union{KeyedDistribution, KeyedSampleable})

Return the keys for the variates of the KeyedDistribution or KeyedSampleable.

KeyedDistributions.distribution — Method

distribution(::KeyedDistribution) -> Distribution
distribution(::KeyedSampleable{F, S, D}) -> D

Return the wrapped distribution.